refactor: move TailwindSwitch to utilities folder
